Add retry code to client and worker
If a Gearman server dies and another server is available, Gearman
clients and workers will automatically try connecting to the next server
in the list.  But a dying server can still cause the next request from a
client or worker to fail.  To help prevent work from being lost, this
revision adds code that will retry connections up to 10 times, using an
exponential back-off algorithm to determine delays between retries.